Verificação antes da exclusão de um plano

+++ app/Http/Controllers/PlanController.php
@@ -42,10 +42,14 @@ class PlanController extends Controller
     }
 
     public function destroy($url){
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository->with('details')->where('url', $url)->first();
 
         if(!$plan){
             return redirect()->back();
+        }
+
+        if($plan->details->count() > 0){
+            return redirect()->back()->with('error', 'Não foi possível excluir este plano. Existem detalhes vinculados à ele.');
         }
 
         $plan->delete();
